gpu: generic: deconvolution: move implementation from ocl folder
